Migrate to webpacker
