+++ .gitignore
@@ -18,3 +18,6 @@ dump.rdb
 # Ignore Byebug command history file.
 .byebug_history
 /ssl/
+/public/packs
+/public/packs-test
+/node_modules
