+++ .postcssrc.yml
@@ -1,0 +1,3 @@
+plugins:
+  postcss-import: {}
+  postcss-cssnext: {}
