--- app/assets/javascripts/application.js
@@ -1,24 +1,0 @@
-// This is a manifest file that'll be compiled into application.js, which will include all the files
-// listed below.
-//
-// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
-// or any plugin's vendor/assets/javascripts directory can be referenced here using a relative path.
-//
-// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// compiled file. JavaScript code in this file should be added after the last require_* statement.
-//
-// Read Sprockets README (https://github.com/rails/sprockets#sprockets-directives) for details
-// about supported directives.
-//
-//= require fetch
-//= require clipboard
-//= require classie
-//= require u2f-api
-//= require vanilla-ujs
-//= require_tree .
-
-window.onload = function () {
-  // Turn all elements with class .copy-button into clipboard.js objects
-  var copyButtons = document.querySelectorAll('.copy-button')
-  var clipboard = new Clipboard(copyButtons)
-}
