+++ app/javascript/copyButton.js
@@ -1,0 +1,7 @@
+import Clipboard from 'clipboard';
+
+document.addEventListener('DOMContentLoaded', function() {
+  // Turn all elements with class .copy-button into clipboard.js objects
+  var copyButtons = document.querySelectorAll('.copy-button')
+  var clipboard = new Clipboard(copyButtons)
+});
