+++ app/javascript/packs/application.js
@@ -1,0 +1,11 @@
+import 'analytics';
+import 'modal';
+import 'copyButton';
+import 'utils/request';
+import 'u2f/registrationPage';
+import 'u2f/authenticationPage';
+import 'legacyGlobals';
+import Rails from 'rails-ujs';
+
+Rails.start();
+
