+++ app/javascript/packs/static.js
@@ -1,0 +1,1 @@
+import 'analytics';
