+++ app/views/layouts/static.html.slim
@@ -7,7 +7,7 @@ html
     meta name="viewport" content="width=device-width, initial-scale=1"
     title= t("shared.app_title")
     = stylesheet_link_tag('application', media: 'all')
-    = javascript_include_tag('application')
+    = javascript_pack_tag('static')
     link rel="icon" type="image/png" href=image_path("favicon.ico")
 
   body data-action=params[:action] data-controller=params[:controller]
