+++ app/views/publishers/contact_info.html.slim
@@ -1,37 +1,36 @@
 javascript:
   (function() {
-
+    'use strict';
     window.addEventListener('load', function() {
       var updateContactInfo = document.getElementById('update_contact_info');
 
       updateContactInfo.addEventListener('submit', function(event) {
         event.preventDefault();
-        spinner.show();
-        submitForm('update_contact_info', 'PATCH')
+        window.spinner.show();
+        window.submitForm('update_contact_info', 'PATCH')
           .then(function() {
-            return pollUntilSuccess('/publishers/domain_status', 2000, 1000, 10);
+            return window.pollUntilSuccess('/publishers/domain_status', 2000, 1000, 10);
           })
           .then(function(response) {
             return response.json();
           })
           .then(function(json) {
-            spinner.hide();
+            window.spinner.hide();
             if (json.error) {
-              flash.clear();
-              flash.append('warning', 'There were errors saving your request:');
-              flash.append('warning', json.error);
+              window.flash.clear();
+              window.flash.append('warning', 'There were errors saving your request:');
+              window.flash.append('warning', json.error);
             } else {
               window.location.href = json.next_step;
             }
           })
           .catch(function(e) {
-            spinner.hide();
-            flash.clear();
-            flash.append('warning', 'An unexpected error occurred saving your changes.');
+            window.spinner.hide();
+            window.flash.clear();
+            window.flash.append('warning', 'An unexpected error occurred saving your changes.');
           });
       }, false);
     }, false);
-
   })();
 
 - content_for(:navbar_content) do
