+++ config/webpack/development.js
@@ -1,0 +1,3 @@
+const environment = require('./environment')
+
+module.exports = environment.toWebpackConfig()
