+++ config/webpack/environment.js
@@ -1,0 +1,3 @@
+const { environment } = require('@rails/webpacker')
+
+module.exports = environment
